requirements.txt
openai==1.66.5
python-dotenv==1.0.1
rich==13.9.4
